requirements.txt
Django==4.0.6
djangorestframework==3.14.0
django-cors-headers==4.2.0
Pillow==10.0.0
coreapi==2.3.3